pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "langfuse>=3.12.1",
    "litellm==1.67.2",
    "openinference-instrumentation-google-adk>=0.1.8",
    "opentelemetry-exporter-otlp>=1.33.0",
    "opentelemetry-sdk>=1.33.0",
    "uvicorn>=0.34.2",
]
